Read setttings in SynchronizationService
